+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Administrator Dark Mode CSS would also partially apply in Light Mode as well if the setting was Auto instead of Always
 * Concurrent handling of subscription_created and subscription_payment_succeeded callbacks leads to inactive subscriptions 
 * Content plugin aslink incompatible with Joomla 4
+* Some system plugins could fail depending on execution order
 
 # 7.1.0